Moved FrenetFrames calculation from TubeBufferGeometry to Curve
