+++ docs/api/extras/core/Curve.html
@@ -60,6 +60,10 @@
 		<h3>[method:Vector getTangentAt]( u )</h3>
 		<div>Returns tangent at equidistant point u on the curve</div>
 
+		<h3>[method:Object computeFrenetFrames]( segments, closed )</h3>
+		<div>
+		Generates the Frenet Frames. Used in geometries like [page:TubeGeometry] or [page:ExtrudeGeometry].
+		</div>
 
 		<h2>Source</h2>
