+++ docs/api/geometries/TubeBufferGeometry.html
@@ -97,21 +97,6 @@
 		An array of [page:Vector3] binormals
 		</div>
 
-
-		<h2>Methods</h2>
-
-
-		<h3>THREE.TubeBufferGeometry.FrenetFrames([page:Curve path], [page:Integer segments], [page:Boolean closed])</h3>
-		<div>
-		path — A path that inherits from the [page:Curve] base class <br />
-		segments — The number of segments that make up the tube <br />
-		closed — Is the tube open or closed
-		</div>
-		<div>
-		A static method that generates the Frenet Frames. This is internally run on any new TubeBufferGeometry and then the
-		generated tangents, normals, and binormals are exposed as properties on the TubeBufferGeometry object.
-		</div>
-
 		<h2>Source</h2>
 
 		[link:https://github.com/mrdoob/three.js/blob/master/src/[path].js src/[path].js]
